requirements.txt
requests==2.32.3
python-dotenv==1.1.0
pandas==2.1.4
pyarrow==20.0.0
ibm-watsonx-ai==1.3.24
ibm-watson-openscale==3.0.47